Added eslint (Airbnb style) config, and laid out better file structure
