+++ .eslintrc.json
@@ -1,0 +1,11 @@
+{
+    "env": {
+        "node": true,
+        "browser": true,
+        "es6": true
+    }, 
+    "parser": "babel-eslint",
+    "extends": "airbnb",
+    "rules": 
+      { "react/jsx-filename-extension": [ 1, { "extensions": [ ".js", ".jsx" ] } ] }
+}
