+++ .expo/packager-info.json
@@ -1,0 +1,9 @@
+{
+  "devToolsPort": 19002,
+  "expoServerPort": 19000,
+  "packagerPort": 19001,
+  "packagerPid": 8400,
+  "expoServerNgrokUrl": "https://3b-rix.virtualsean.storiesui.exp.direct",
+  "packagerNgrokUrl": "https://packager.3b-rix.virtualsean.storiesui.exp.direct",
+  "ngrokPid": 8441
+}
